♻️ Splitted test for Eurostat on source tests and task tests
